Move `main.rs` into `lib.rs`.
This PR changes the way Rust compiles the canister from a binary into
a library. The outcome is exactly the same, but moving `main.rs` into
`lib.rs` gives us some additional advantages:

1. Tests written can now use methods/tools that are only available in
   `cfg(test)` environments. This wasn't possible with `main.rs` as it
   was considered to be outside of the canister's crate.

2. Detection of unused code. Because methods now need not be public,
  clippy can warn us of any unused code.

Note that I relaxed the clippy check as to allow warnings for now until
these new warnings are resolved in additional PRs.
