+++ .github/workflows/workflow.yml
@@ -35,7 +35,7 @@ jobs:
       - uses: actions-rs/clippy-check@v1
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
-          args: --all-targets --all-features -- -D warnings
+          args: --all-targets --all-features
 
       - name: Run Tests
         shell: bash
